new level-1  page started

+++ server/controllers/gradeSectionController.js
@@ -2,6 +2,7 @@ const GradeSection = require('../models/grade-sections');
 const ClassificationGrade = require('../models/classification-grade');
 const SectionClass = require("../models/section-class");
 const GradeSubject = require('../models/grade-subject');
+const StudentGrade = require("../models/student-grade");
 
 // Controller methods
 const GradeSectionController = {
@@ -9,23 +10,25 @@ const GradeSectionController = {
     createGradeSection: async (req, res) => {
         try {
             const { classification_grade, section } = req.body;
+            const class_grade = await ClassificationGrade.findById(classification_grade);
+            if (!class_grade) {
+                return res.status(404).json({ message: 'Classification grade not found' });
+            }
             const gradeSection = new GradeSection({ classification_grade, section });
             const savedGradeSection = await gradeSection.save();
-            
-            const cgrade = await ClassificationGrade.findById(classification_grade);
-            const subjectGrades = await GradeSubject.find({ curriculum_grade: cgrade.curriculum_grade}); 
-            
+
+            const subjectGrades = await GradeSubject.find({ curriculum_grade: class_grade.curriculum_grade });
             for (const subjectGrade of subjectGrades) {
                 const newSubjectClass = new SectionClass({
                     grade_section: savedGradeSection._id,
                     grade_subject: subjectGrade._id
                 });
                 await newSubjectClass.save();
-            }   
+            }
             res.status(201).json(savedGradeSection);
         } catch (error) {
-            res.status(400).json({ message: error.message });
-            console.log(error);
+            res.status(500).json({ message: error.message });
+            //console.log(error);
         }
     },
 
@@ -39,15 +42,22 @@ const GradeSectionController = {
         }
     },
 
-
-
     // Delete a gradeSection by ID
     deleteGradeSection: async (req, res) => {
         try {
             const { id } = req.params;
+            const isReferenced = await StudentGrade.findOne({ grade_section: id });
+
+            if (isReferenced) {
+                return res.status(400).json({
+                    success: false,
+                    message: 'Cannot delete GradeSection because it is referenced in StudentGrade.',
+                });
+            }
+            // If not referenced, proceed with deletion    
             const deletedGradeSection = await GradeSection.findByIdAndDelete(id);
             if (!deletedGradeSection) {
-                return res.status(404).json({ message: 'Section not found' });
+                return res.status(404).json({ message: 'Grade Section not found' });
             }
             res.status(200).json({ message: 'Section deleted successfully' });
         } catch (error) {
